ISO: Split out test cases
