+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/LibraryInitializerTest.kt
@@ -1,7 +1,9 @@
 package at.asitplus.wallet.lib
 
 import at.asitplus.iso.IssuerSignedItem
-import at.asitplus.iso.IssuerSignedItemSerializer
+import at.asitplus.iso.IssuerSignedList
+import at.asitplus.iso.IssuerSignedListSerializer
+import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.testballoon.invoke
 import at.asitplus.wallet.lib.data.AttributeIndex
@@ -101,21 +103,28 @@ val LibraryInitializerTest by testSuite {
         )
 
         JsonCredentialSerializer.encode(MockIssuerSignedValue("encoded")) shouldBe
-            vckJsonSerializer.encodeToJsonElement(MockIssuerSignedValue("encoded"))
+                vckJsonSerializer.encodeToJsonElement(MockIssuerSignedValue("encoded"))
 
-        val item = IssuerSignedItem(
-            digestId = 1u,
-            random = Random.nextBytes(16),
-            elementIdentifier = elementId,
-            elementValue = MockIssuerSignedValue("round-trip"),
+        val list = IssuerSignedList(
+            listOf(
+                ByteStringWrapper(
+                    IssuerSignedItem(
+                        digestId = 1u,
+                        random = Random.nextBytes(16),
+                        elementIdentifier = elementId,
+                        elementValue = MockIssuerSignedValue("round-trip"),
+                    )
+                )
+            )
         )
-        val encodedItem =
-            coseCompliantSerializer.encodeToByteArray(IssuerSignedItemSerializer(isoNamespace, elementId), item)
-        val decodedItem = coseCompliantSerializer.decodeFromByteArray(
-            IssuerSignedItemSerializer(isoNamespace, elementId),
-            encodedItem
+        val encodedList = coseCompliantSerializer.encodeToByteArray(
+            IssuerSignedListSerializer(isoNamespace),
+            list
         )
-
-        decodedItem shouldBe item
+        val decodedList = coseCompliantSerializer.decodeFromByteArray(
+            IssuerSignedListSerializer(isoNamespace),
+            encodedList
+        )
+        decodedList shouldBe list
     }
 }
